WIP: Restructure API, try to fix race condition
There's a race condition when the peer is started and the handlers are
registered. The peer needs to run when it is passed to the handler. The
handler will likely register callbacks for onSubscription and
onAnnouncement. But it may also try to make announcements or
subscriptions itself, which requires the peer to run. But when the peer
starts running before the callbacks are registered, it may miss some
subscriptions or announcements.